Changed to working fork of ncurses wrapper
The [original version](https://github.com/rnowley/nim-ncurses/) has an out of date .nimble file which does not work with the current version of nimble.
See [this issue] in the original repo. Someone made a [PR](https://github.com/rnowley/nim-ncurses/pull/25) with a fix, but the maintainer seems inactive.

I forked the repo, fixed the .nimble file and corrected a compiler error within ncurses.nim.