clean up status packages
These packages were removed, abandoned or reorganised years ago and
should see no usage in the community regardless most of the code was
simply moved elsewhere long ago, before the code reached a useful level
of quality.

They are unmaintained and probably don't work any more, so keeping them
in packages both useless and misleading.

If there was a feature to remove them from the default search function
of `nimble.directory`, that could be used in the future for cases like
this so as not to break old links.